Refactor backend data_classes code
