+++ digiprod_gen/backend/image/caption.py
@@ -5,7 +5,7 @@ from typing import List
 from PIL import Image
 from digiprod_gen.backend.image.conversion import pil2bytes_io, bytes2pil
 from digiprod_gen.backend_api.models.mba import MBAProduct
-from digiprod_gen.backend.data_classes.common import ImageCaptioningModel
+from digiprod_gen.backend_api.models.common import ImageCaptioningModel
 from digiprod_gen.backend_api.utils import booleanize
 
 def image2prompt(img_pil: Image) -> str:
