+++ digiprod_gen/frontend/session.py
@@ -5,9 +5,9 @@ import streamlit as st
 from digiprod_gen.backend_api.browser.crawling.utils.utils_mba import get_random_headers
 from digiprod_gen.backend_api.browser.selenium_fns import SeleniumBrowser
 from digiprod_gen.backend_api.models.mba import CrawlingMBARequest, MBAMarketplaceDomain, MBAProductCategory
-from digiprod_gen.backend.data_classes.session import SessionState, ImageGenData, CrawlingData, MBAUploadData, \
+from digiprod_gen.backend_api.models.session import SessionState, ImageGenData, CrawlingData, MBAUploadData, \
     DigiProdGenStatus, DigitProdGenViews
-from digiprod_gen.backend.data_classes.config import DigiProdGenConfig
+from digiprod_gen.backend_api.models.config import DigiProdGenConfig
 from digiprod_gen.frontend.backend_caller import BackendCaller
 from digiprod_gen.backend_api.utils import is_debug, delete_files_in_path
 from digiprod_gen.backend_api.transform.transform_fns import request2mba_overview_url
