+++ digiprod_gen/frontend/tab/image_generation/prompt_generation.py
@@ -5,10 +5,9 @@ from streamlit.delta_generator import DeltaGenerator
 from langchain.chat_models import ChatOpenAI
 
 from digiprod_gen.backend.generative_ai.text.text_gen_fns import get_midjourney_prompt_gen, mba_products2llm_prompt_gen_input
-from digiprod_gen.backend.prompt_engineering.utils import extract_list_from_output
-from digiprod_gen.backend.data_classes.common import MBAMidjourneyOutputModel
+from digiprod_gen.backend_api.models.common import MBAMidjourneyOutputModel
 from digiprod_gen.frontend.session import read_session
-from digiprod_gen.backend.data_classes.session import SessionState
+from digiprod_gen.backend_api.models.session import SessionState
 
 
 def prompt_generation(st_tab_ig: DeltaGenerator):
