+++ digiprod_gen/frontend/tab/image_generation/selected_products.py
@@ -3,7 +3,7 @@ from streamlit.delta_generator import DeltaGenerator
 
 import streamlit as st
 
-from digiprod_gen.backend.data_classes.session import CrawlingData, SessionState
+from digiprod_gen.backend_api.models.session import CrawlingData, SessionState
 from digiprod_gen.backend.io.io_fns import image_url2image_bytes_io
 from digiprod_gen.backend_api.utils import split_list
 from digiprod_gen.backend_api.models.mba import MBAProduct
